pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "ai_credentials"
version = "0.1.0"
description = "A Python library for managing credentials with Rust backend"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
]
keywords = ["credentials", "security", "password-manager"]

dependencies = [
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "maturin"]

[project.urls]
Homepage = "https://github.com/yourusername/ai_credentials"
Repository = "https://github.com/yourusername/ai_credentials.git"

[tool.maturin]
features = ["pyo3/extension-module"]
compatibility = "linux" 